# openapi.yaml
openapi: 3.0.3
info:
  title: go-drink API
  description: Go buy something to drink!
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
  /items:
    get:
      description: Retrieve a list of all available items
      responses:
        500:
          $ref: "#/components/responses/500-empty-array"
        200:
          description: a list of all available Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/item"
...
  /users:
    get:
      description: Return a list of all registered users in the application
      responses:
        500:
          $ref: "#/components/responses/500-empty-array"
        401:
          $ref: "#/components/responses/401"
        200:
          description: A list of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
...
components:
  responses:
...
    401:
      description: If the action requires a higher authorization / authentication, the response is empty
...
    500-empty-array:
      description: In case of an error, an empty list with no further information is returned
      content:
        application/json:
          schema:
            type: array
  schemas:
    item:
      type: object
      description: an item that can be bought
      properties:
        id:
          type: string
          description: uuid v4
        name:
          type: string
          description: a unique name, no longer than 64 bytes
        price:
          type: integer
          description: price in cents
        image:
          type: string
          description: an image in form of a base64 encoded string. Not larger than 2 MiB
        amount:
          type: integer
          description: the amount of the available items. Can't be negative
        barcode:
          type: string
          description: a barcode identifying the item
    user:
      type: object
      description: A user that can authenticate in some way to the application
      properties:
        id:
          type: string
          description: uuid v4
        username:
          type: string
          description: A username that can be used to identify a user
        email:
          type: string
          description: An email address that might be used to contact that user
        role:
          type: string
          description: The role a user has in the application. Can be one of "user" od "admin"
        credit:
          type: integer
          description: The amount of money a user can spend on items
...